asgiserver/requirements.txt
Django>=3.0.6,<3.1
uWSGI>=2.0.18,<2.1
psycopg2-binary>=2.8
netmiko
channels==2.4.0
channels_redis